% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_displacement_from_start.R
\name{add_displacement_from_start}
\alias{add_displacement_from_start}
\title{Add displacement from the starting position}
\usage{
add_displacement_from_start(
  tracks_movement,
  x_col = "x",
  y_col = "y",
  id_col = "track_id",
  frame_col = "frame",
  pixel_size_mm = 10/54
)
}
\arguments{
\item{tracks_movement}{A data frame containing tracking data.}

\item{x_col}{Name of the x-coordinate column. Default is "x".}

\item{y_col}{Name of the y-coordinate column. Default is "y".}

\item{id_col}{Name of the track ID column. Default is "track_id".}

\item{frame_col}{Name of the frame column. Default is "frame".}

\item{pixel_size_mm}{Pixel size in millimetres. Default is 10 / 54.}
}
\value{
A data frame with added columns for starting x and y positions,
displacement from start in pixels, and displacement from start in millimetres.
}
\description{
Calculates the displacement of each tracked object from its initial
x and y position. Displacement is returned in both pixels and millimetres.
}
