% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibrate_scale.R
\name{calibrate_scale}
\alias{calibrate_scale}
\title{Calibrate pixel size from a known distance}
\usage{
calibrate_scale(frames, frame_number = 1, known_distance_mm)
}
\arguments{
\item{frames}{A list of image frames, usually loaded from a video.}

\item{frame_number}{Numeric. The frame number to use for calibration.
Default is \code{1}.}

\item{known_distance_mm}{Numeric. The real-world distance between the two
clicked points, in millimetres.}
}
\value{
A numeric value representing pixel size in mm/pixel.
}
\description{
Displays one video frame and allows the user to click two points with a known
real-world distance. The function calculates the pixel size in millimetres
per pixel.
}
\examples{
\dontrun{
pixel_size_mm <- calibrate_scale(
  frames,
  frame_number = 1,
  known_distance_mm = 10
)
}

}
