% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crop_patch.R
\name{crop_patch}
\alias{crop_patch}
\title{Crop a small image patch around a coordinate}
\usage{
crop_patch(mat, x, y, radius = 10)
}
\arguments{
\item{mat}{A numeric matrix, usually a grayscale image matrix.}

\item{x}{Numeric. X coordinate of the patch centre.}

\item{y}{Numeric. Y coordinate of the patch centre.}

\item{radius}{Numeric. Radius of the patch in pixels. Default is \code{10}.}
}
\value{
A numeric matrix containing the cropped image patch.
}
\description{
Crops a square patch from a matrix around a given x/y coordinate. The crop is
automatically limited to the matrix boundaries.
}
\keyword{internal}
