% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_frame.R
\name{plot_frame}
\alias{plot_frame}
\title{Plot a single video frame}
\usage{
plot_frame(frames, frame_number = 1)
}
\arguments{
\item{frames}{A list of image frames, usually loaded from a video.}

\item{frame_number}{Numeric. The frame number to display. Default is \code{1}.}
}
\value{
Invisibly returns \code{NULL}. This function is used for plotting.
}
\description{
Displays one frame from a list of image frames. If the frame has multiple
channels, it is converted to grayscale before plotting.
}
\examples{
\dontrun{
plot_frame(frames, frame_number = 1)
}

}
