% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_total_distance.R
\name{plot_total_distance}
\alias{plot_total_distance}
\title{Plot cumulative distance travelled over time}
\usage{
plot_total_distance(tracks_movement)
}
\arguments{
\item{tracks_movement}{A data frame containing movement data. It should
contain \code{track_id}, \code{time_s}, and \code{total_distance_mm} columns.}
}
\value{
A \code{ggplot} object showing cumulative distance over time.
}
\description{
Creates a line plot showing the cumulative distance travelled by each tracked
object over time.
}
\examples{
\dontrun{
plot_total_distance(tracks_movement)
}

}
