% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_ROI.R
\name{select_roi}
\alias{select_roi}
\title{Select a rectangular region of interest from a video frame}
\usage{
select_roi(frames, frame_number = 1)
}
\arguments{
\item{frames}{A list of image frames, usually loaded from a video.}

\item{frame_number}{Numeric. The frame number to display for ROI selection.
Default is \code{1}.}
}
\value{
A list containing \code{x_min}, \code{x_max}, \code{y_min}, and \code{y_max} coordinates
of the selected ROI.
}
\description{
Displays one frame from a list of video frames and allows the user to
interactively select a rectangular region of interest (ROI) by clicking two
opposite corners. The function returns the ROI coordinates as a list.
}
\examples{
\dontrun{
roi <- select_roi(frames, frame_number = 1)
}

}
